+++ changelog.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Changed
 - Upgrade dependencies
+- Add `jsx-runtime` config to react config
 
 ## [0.3.0] - 2023-03-25
 ### Changed